Fixes and tweaks 3
* Fix soilmodel bug (showing wrong soilmodel when switching between projects)
