+++ src/scenes/t03/actions/actions.tsx
@@ -1,29 +1,24 @@
-import {UPDATE_BOUNDARIES} from '../reducers/boundaries';
-import {START_CALCULATION, UPDATE_CALCULATION} from '../reducers/calculation';
-import {CLEAR, UPDATE_MODEL, UPDATE_MT3DMS, UPDATE_STRESSPERIODS} from '../reducers/model';
-import {UPDATE_OPTIMIZATION} from '../reducers/optimization';
-import {
-    ADD_SOILMODEL_LAYER, ADD_ZONE, CLONE_SOILMODEL_LAYER, CLONE_ZONE,
-    REMOVE_SOILMODEL_LAYER, REMOVE_ZONE, UPDATE_SOILMODEL, UPDATE_SOILMODEL_LAYER,
-    UPDATE_SOILMODEL_RELATIONS, UPDATE_ZONE
-} from '../reducers/soilmodel';
-
 import FlopyPackages from '../../../core/model/flopy/packages/FlopyPackages';
-import {FlopyMt3d} from '../../../core/model/flopy/packages/mt';
 import {IMessage} from '../../../core/model/messages/Message.type';
-import {Calculation, ModflowModel, Stressperiods, Transport, VariableDensity} from '../../../core/model/modflow';
+import {Calculation, ModflowModel, Transport, VariableDensity} from '../../../core/model/modflow';
 import {BoundaryCollection} from '../../../core/model/modflow/boundaries';
 import {Optimization} from '../../../core/model/modflow/optimization';
 import {Soilmodel, SoilmodelLayer} from '../../../core/model/modflow/soilmodel';
 import {Zone} from '../../../core/model/modflow/soilmodel';
-import LayerParameterZonesCollection from '../../../core/model/modflow/soilmodel/LayerParameterZonesCollection';
+import {UPDATE_BOUNDARIES} from '../reducers/boundaries';
+import {START_CALCULATION, UPDATE_CALCULATION} from '../reducers/calculation';
 import {ADD_MESSAGE, REMOVE_MESSAGE, UPDATE_MESSAGE} from '../reducers/messages';
+import {CLEAR, UPDATE_MODEL} from '../reducers/model';
+import {UPDATE_OPTIMIZATION} from '../reducers/optimization';
 import {
     PROCESSING_PACKAGES,
-    RECALCULATE_PACKAGES,
     UPDATE_PACKAGES,
     UPDATE_PROCESSED_PACKAGES
 } from '../reducers/packages';
+import {
+    ADD_SOILMODEL_LAYER, ADD_ZONE, REMOVE_SOILMODEL_LAYER, REMOVE_ZONE, UPDATE_SOILMODEL, UPDATE_SOILMODEL_LAYER,
+    UPDATE_ZONE
+} from '../reducers/soilmodel';
 import {UPDATE_TRANSPORT} from '../reducers/transport';
 import {UPDATE_VARIABLE_DENSITY} from '../reducers/variableDensity';
 
@@ -37,13 +32,6 @@ export function updateModel(modflowModel: ModflowModel) {
     return {
         type: UPDATE_MODEL,
         model: modflowModel.toObject()
-    };
-}
-
-export function updateStressperiods(stressperiods: Stressperiods) {
-    return {
-        type: UPDATE_STRESSPERIODS,
-        payload: stressperiods.toObject()
     };
 }
 
@@ -96,31 +84,10 @@ export function updateProcessedPackages(packages: FlopyPackages) {
     };
 }
 
-export function recalculatePackages() {
-    return {
-        type: RECALCULATE_PACKAGES
-    };
-}
-
-export function updateMt3dms(mt3dms: FlopyMt3d) {
-    return {
-        type: UPDATE_MT3DMS,
-        payload: mt3dms.toObject()
-    };
-}
-
 export function addLayer(layer: SoilmodelLayer) {
     return {
         type: ADD_SOILMODEL_LAYER,
         layer: layer.toObject()
-    };
-}
-
-export function cloneLayer(layerId: string, newLayerId: string) {
-    return {
-        type: CLONE_SOILMODEL_LAYER,
-        layer_id: layerId,
-        new_layer_id: newLayerId
     };
 }
 
@@ -135,13 +102,6 @@ export function addZone(zone: Zone) {
     return {
         type: ADD_ZONE,
         zone: zone.toObject()
-    };
-}
-
-export function cloneZone(zoneId: string, newZoneId: string) {
-    return {
-        type: CLONE_ZONE,
-        zone_id: zoneId, new_zone_id: newZoneId
     };
 }
 
@@ -170,13 +130,6 @@ export function updateSoilmodel(soilmodel: Soilmodel) {
     return {
         type: UPDATE_SOILMODEL,
         soilmodel: soilmodel.toObject()
-    };
-}
-
-export function updateSoilmodelRelations(relations: LayerParameterZonesCollection) {
-    return {
-        type: UPDATE_SOILMODEL_RELATIONS,
-        relations: relations.toObject()
     };
 }
